post_notes

Module settings

Structs

Args 🔒

Command line arguments - mirrors Settings structure.

CliFrontMatterSettings 🔒

Optional front matter settings used to parse command line arguments - similar to FrontMatterSettings.

CliPathSettings 🔒

Optional path settings used to parse command line arguments - mirros PathSettings.

Field 🔒

Represents a front matter field holding data of a certain ValueType.

FrontMatterSettings

All settings that can be configured regarding the parsing of the front matter.

PathSettings

All settings that can be cofnigured regarding the directories which will be referenced during the site generation.

PipelineSettings

Contains settings related to the build pipeline.

PipelineStep

A single step in the build pipeline.

Schema

Represents the schema of the front matter.

Settings

Configurable application settings which get derived from command line arguments and the Config.toml.

Enums

SchemaValue

A wrapper around Schema, that denotes the origin of the schema.

ValueType 🔒

Represents the type of value the front matter field holds.

Constants

CONFIG_PATH 🔒

DEFAULT_ASSET_PATH 🔒

DEFAULT_INPUT_PATH 🔒

DEFAULT_OUTPUT_PATH 🔒

DEFAULT_TEMPLATE_PATH 🔒

DEFAULT_VOLATILE_PATH 🔒

Functions

get_settings

Loads the configured settings from either Config.toml or the command line arguments.

merge_settings 🔒

Read Settings from Config.toml or command line arguments.