post_notes

Module settings

Structs

Args 🔒

Command line arguments - mirrors Settings structure.

CliPathSettings 🔒

Optional path settings used to parse command line arguments - mirros PathSettings.

PathSettings

All settings that can be cofnigured regarding the directories which will be referenced during the site generation.

Settings

Configurable application settings which get derived from command line arguments and the Config.toml.

Constants

CONFIG_PATH 🔒

DEFAULT_ASSET_PATH 🔒

DEFAULT_INPUT_PATH 🔒

DEFAULT_OUTPUT_PATH 🔒

DEFAULT_TEMPLATE_PATH 🔒

Functions

get_settings

Loads the configured settings from either Config.toml or the command line arguments.

merge_settings 🔒

Read Settings from Config.toml or command line arguments.